face_reception/app/requirements.txt
psutil
psycopg2-binary
requests
requests-toolbelt
pillow
uvicorn[standard]
fastapi==0.98.0
python-multipart
